refactor(mobile): extract ReaderShell, route public reader through it
ReaderShell owns the shared reader body (WebView, vocab+inline-translation,
highlights, selection, bars, top bar, footer, sheets, scroll restore, message
routing), parameterized by a {kind:'edition'|'userbook'} source. The public
reader becomes a thin wrapper that owns only its catalog data hooks. Behavior
unchanged for public; user-book route migrates next. tsc clean.
